Made getBranches and getTags in Repository more sane
Previously, they did very different things, which were quite confusing and
inconsistent. Now there is a helper method, listRefs, that takes a
subdirectory, and will list all refs, whether packed or unpacked, that are in
that subdirectory.

getBranches is now just listRefs("heads"), getTags is listRefs("tags"), and
there is a new method, listAllRefs, that lists everything in the refs/ folder.

Additionally, all refs returned are normalized - that is, they have the full
refs/whatever, instead of just being short branchnames.

Signed-off-by: Dave Watson <dwatson@mimvista.com>