IndexTreeWalker walks index and tree in parallel
This class, along with the IndexTreeVisitor, is intended to solve
the problem of how to compare members of the index and a tree without
turning the index into a tree, which will waste memory and probably time.

This will be used by the ResetOperation, BranchOperation, WorkdirCheckout, and
IndexDiff, to do things more efficiently, in a single pass instead of several.

Signed-off-by: Dave Watson <dwatson@mimvista.com>