Adds ability to perform a git-reset on repository
There is now a ResetAction/ResetOperation, available from the Team
menu at the project level. You can perform soft/mixed/hard resets to
any local or remote branch or tag. At this point you can't specify
an arbitrary commit, only one that is pointed to by a branch or tag,
but the underlying code really doesn't care either way, so it just
needs more UI support.

The ResetOperation updates the current branch, working directory, and
index (depending on reset type), just like git-reset. It currently
operates on the entire repository - there is not currently the option
to just do it to a subset.

Includes the BranchSelectionDialog, which lists all available refs,
and allows the user to select one. This has an optional reset area,
with soft/mixed/hard options.

Signed-off-by: Dave Watson <dwatson@mimvista.com>