Created WorkDirCheckout class
This class can take a GitIndex and either one or two trees.
Depending on the setting, it will either throw a CheckoutConflictException,
or silently deal with the conflicts. This includes tests for directory/file
conflicts.

It performs the same basic tasks as git-read-tree, and there are test
cases included to test many different situations, including directory/
file conflicts, and other type of conflicts.

Signed-off-by: Dave Watson <dwatson@mimvista.com>