Add feature and associated plugin.
I made a feature and associated plugin for egit. The feature will allow
you guys to publish and update site and the plugin adds an entry for the
Egit in Help -> About Eclipse -> Feature Details. This is useful as it
lets users easily disable the plugins if they want to for some reason.

Signed-off-by: Ben Konrath <bkonrath@redhat.com>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>