Add javadoc to undocumented public methods, variables and classes.
Force javadoc comments on public and protected methods.
Change visibility on a few methods to avioid forcing javadoc
comments on private and package visible methods and vairables.

This doesn't mean that there should not be any commnents in there, but
I think it is wise to up the standards and enforce it.

Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>