Ignore warning for "Access to enclosing method/field X emulated by a synthetic accessor method"
There is no measurable performance degradation from this. Increasing visibility
of methods and fields in order to get rid of this warning reduces encapsulation.

Instead of choosing the Ignore setting we remove the setting from the project,
which may be somwhat hackish but has the benefit that the setting can be changed
in the workspace for those that wants to investigate the issue.

Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>