Git clone Wizard
This wizard allows you to clone locally, using the git
protocol directly or using SSH.

Local cloning is done by spawning a native git process, almost
identical to cloning over the network. This gives us a way
to test the git protocol more and save the time to implement
a real shortcut for now.

SSH is username/password only. Keys need more management, so
we will get back to that.

Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>
Signed-off-by: Roger C. Soares <rogersoares@intelinet.com.br>