CompareEditor: Show diffs for changed files under a new or deleted directory.
When a directory is added or removed in a commit the files under it
were not being shown in the structure compare viewer. This fixes it by
adding diff nodes for all the files under this directory.

This patch also fixes some files showing as being removed in the
structure compare when they were actually being added.

Signed-off-by: Roger C. Soares <rogersoares@intelinet.com.br>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>