Implement a Tree iterator
This is an implementation of the Iterator interace. It can
be used for iterating over all entries in a tree in Git
order. Since it is an iterator it can be used to walk more
than one tree in parallel.

Although only post order is needed, preorder and leaves-only
mode is also implemented here.

Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>