Refactor index reading code from PackFile to PackIndex
This makes it easier for us to construct different implementations
of the packfile index code, based on which version of the index we
need to read.  Index v2 is quite a bit different in layout from the
index v1 format and we may want a different implementation to access
its contents.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>