Read pack index files using FileInputStream and not WindowedFile
Since we no longer mmap in the pack .idx file but instead read the
entire thing (as that turned out to be faster) we don't want any
of the overhead associated with setting up the window cache or a
file read into there.

Instead we setup a FileInputStream and try to perform very large
reads from the file descriptor.  There is no point in using any
buffering here as we issue no more than 257 reads of the index
file.  (1 for the fan-out table, 256 for each of the levels of the
fan-out table).

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>