Move index:pack object count checking into PackFile class
If we ever start to delay the opening of the packfile until we
know we actually want/need object data from within the pack we
are likely to not be able to pass an actual length into our
PackIndex class when we create the index object.  Consequently
we need to defer this test of the object counts until we have
both open and know they are opened successfully.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>