Refactor PackIndex version 1 handling code into PackIndexV1
Setting up an abstract base class of PackIndex allows us to create
two different implementations of this API, one for the legacy v1
packfile index format and one for the newer v2 format.  We want
to use two completely different implementations here because we
can then continue to run the on-disk format in memory, which will
reduce our overall loading costs.

Most projects will use only one or the other format for all of
their packfiles, and not too long from now that will be v2.  A
good JIT should be able to optimize most calls into the v2 path
without virtual dispatches, so we get some nice flexiblity for
almost zero cost.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>