Add complete reading support for pack index v2
Since C Git has had support for pack index v2 for quite a while now
we are starting to see a lot of repositories in the wild using this
as their native packfile index format.  At some point in the near
future this will also become the default format that C Git creates,
so we must support reading from it in jgit.

At present we ignore the CRC32 table within the index file as we do
not perform any repacking actions.  Even if we do implement a repack
we won't need the CRC32 data unless we are actively doing a repack,
which is few and far between object reads.  So it makes to avoid this
data when we load the index into memory.

Like the v1 format we load the entire .idx file into memory and try
to break it out by fan-out structures.  However we keep the on disk
format as much as we can, improving cache-line locality during the
binary search for an object name.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>