Stop creating "new format" loose objects as C Git no longer likes them
The new format of loose objects (aka the pack-like loose object format)
is just not something that went very well in C Git.  It was backed out
and is no longer a format that C Git creates.  However since it was in
the wild in released versions of C Git we must continue to be able to
read the loose object format.  But we should also stop creating it.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>