Don't make unit tests depend upon compressed file lengths
We cannot test for the size of the empty tree object as
its compressed size may differ on different systems.  It
depends highly on how we setup the Deflater and what the
underlying zlib implementation is able to do for us as to
how many bytes the resulting tree file will be.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>