Extract StGitPatch from Repository to its own top-level type
The concept of an StGitPatch and StGit in general is an add-on to
Git.  We probably shouldn't have all of the StGit logic built into
our core Repository class, but instead should try to consolidate
the StGit interface logic into its own package, where it can grow
and do things it needs to do.

According to Git Robin created StGitPatch back in 2007, so I am
putting the copyright header on here and attributing it to him.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>