Refactor RepositoryState into its own top-level type
The Repository class is getting rather long, and we are likely
to need to add further code to it before jgit can be said to
be a "complete" implementation of the Git repository system. As
the state code is rather lengthy in and of itself I am pulling
it all out to its own top-level type, cutting down the overall
size of our Repository class.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>