Refactor PackFile construction to take the idxFile path from scan
C Git ensures packfile safety by putting .pack files into the
$GIT_DIR/objects/pack directory before it puts the corresonding
.idx file into the same directory.  This way running readers do
not attempt to read a packfile for which they do not yet have a
valid index file available for scanning.

This trick is used in C Git because C Git wants to only open a
pack .idx file, saving the actual opening of the .pack file to
be lazy and on-demand when an object is needed from it.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>