Defer opening pack files until we actually need object data
We now try to avoid opening a packfile for reading until we
really need to access object data stored within that packfile.
This improves early startup performance on a repository with
a lot of packfiles, and reduces system resource usage if we
mostly just need to stick with a single packfile, such as
for recent history.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>