Use a single WindowCache for the entire Eclipse workspace
As the JVM and its heap are a shared resource that all projects
loaded and accessed through egit must share we can get better memory
usage if the WindowCache is shared across all repositories, rather
than being repository specific.  This allows a very large project
to use as much memory as possible, without it suddenly dominating
the JVM as access to the other projects can cause the cache to evict
windows from the large project.

Until we actually have egit connected to the Eclipse workspace
preference store we settle for getting our window configuration
from the user's ~/.gitconfig, assuming that any sort of default
that applies to all repositories is sufficient to control the
Eclipse workspace usage.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>