Modify the WindowedFile API to pass around a WindowCursor during reads
By passing a WindowCursor object that is "stack allocated" by the
caller we can store temporary metadata necessary to perform any
read operation, and have that metadata clear out when the caller
returns from their stack frame and pops the only reference to the
WindowCursor object they created for the read operation.

Its not immediately obvious yet but the idea behind this cursor is
to allow storage of both the ByteWindow and the underlying data it
relies upon, but to do so using only strongly held references.  We
can then transition the WindowCache to use SoftReferences to retain
the bulky internal data of a ByteWindow and let the JVM's GC decide
when it needs to make Git drop memory out of the cache because the
available heap space has run low.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>