Change ByteWindow implementations to hold data by SoftReferences
By holding the bulky byte[] or ByteBuffer objects using only a soft
reference when in the WindowCache we can allow the JIT to decide if
it needs to evict our data in order to meet its memory targets.  But
while someone is actively reading from the windowed file we hold the
buffer using strong references, to ensure that the JIT does not clear
a particular window if some thread is trying to copy or inflate data
from within its bounds.

At this point the WindowCache/ByteWindow API is really tightly hooked
together, so I am hiding the entire window type concept within own
package so that application code does not try to use it.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>