Simplify WindowCache and WindowedFile internal APIs to be private
Since a WindowedFile must know the inner details of a WindowCursor,
and those details are not public, I have chosen to reduce the code
we export as public APIs by taking out WindowProvider and folding
it into WindowedFile directly.

This reduces the overall complexity of the code and should make the
window reads a tiny bit faster as we have one less indirection to
get between the PackFile and the actual data array.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>