Use a proper ReferenceQueue to avoid duplpicate window entries
If a window gets cleared by the GC the isEnqueued flag does not get
set by the JVM unless it was supplied a ReferenceQueue when we made
the SoftReference instance.  So we now use our own ReferenceQueue to
list the windows that have been culled by the JVM.

We also avoid opening the first window twice.  If the openCache()
call does a read from the file it may try to access the same data
window that the application was trying to access when the file is
discovered to be closed.  If this happens the openCache call may
have already populated the window we need for us, and we shouldn't
attempt to open it yet a second time.  Such a duplicate mapping is
very wasteful of precious virtual memory resources.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>