Implement core.deltaBaseCacheLimit to improve unpacking performance
When we are unpacking delta chains for revision tree walks it is not
uncommon for the same base tree to pop up several times in fairly
quick succession.  If the delta chain required to get to that tree
is relatively deep (over 5 or 10 items) then it can be expensive to
obtain that tree every time we need data.

We use a direct mapped cache of 256 entries, with an LRU ordering to
limit how much data is stored in the cache at once.  If we get a hash
collision then we clear out the old entry and insert the new one.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>