Precount the number of tree entries during readTree
Rather than growing our tree entry buffer we can count up the
number of entries we need before we start to do any sort of
tree reading, and allocate exactly that many entries for the
result.  This should avoid from overallocating the array as
well as from underallocating (and thus growing) it.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>