Change the ObjectLoader.getType method to return int
Working with the integer type code from the packfile is faster and
easier than working with the String literal.  We actually can take
out a lot of special cases by using the packfile type code for all
of our return values, thus making it easier to read the IO code.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>