Allow RevFilters to break out of the main walk loop
The new StopWalkException is used to abort the revision walker if we
get a commit and we know we cannot continue further.  Currently this
only happens when all commits in the queue are UNINTERESTING, but a
RevFilter may also be able to figure out that there is no point in
walking back further through the graph.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>