Teach FileMode about S_IFGITLINK being 0160000
C Git understands this type of tree entry to be a commit inside of
another Git repository other than the one that contains this tree.

We currently don't support such an idea in jgit but we might as
well provide the enumerated value so support can grow in the future.
The value has been reserved, is in the wild, and people are building
production repositories with it.  We will have to honor this at
some point in the future no matter what happens.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>