Teach FileMode about MISSING being 0
While walking multiple trees in parallel it can be handy to use a
special file mode to mean "this tree lacks an entry for the current
name".  In C Git the mode of 0 is often used for this sort of value,
as it is easy to test and its missing all bits of the S_IFMT mask
so its unlikely a POSIX OS will give us an inode with a mode of 0.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>