Create IndexPack task for jgit
This is a basic routine that takes any packfile on any type of
java.io.InputStream and copies it to a local file as well as
generates the necessary .idx that goes along with it.  Today
it will handle only pack version 2/3 (though only 2 is actually
used in practice) and it will not handle fixing up thin packs.
To fix thin packs we need to copy existing objects from the
repository into the packfile.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>