Evaluate RevFilters before inserting parent commits to pending
By filtering up front we better position ourselves to avoid expensive
parent rewrite related work when the commit is of no interest to the
application.  Currently there is no logic or performance impact to
this change, but it relocates the filter earlier and provides a nicer
structure to insert parent rewriting logic prior to injecting them
into our pending queue.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>