Dispose of a RevCommit as soon as it is no longer interesting
If we are not going to produce a RevCommit out of the pending queue
then there is no interest in its data beyond our need to add all of
its parents into the pending queue.  Once that operation is complete
we can dump the buffer and release the commit's memory.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>