Hand unroll the ObjectId.equals method
The Java compilers and the JIT were really just not doing a very
good job compiling this particular loop.  Its a critical part of the
system as we frequently need to compare two ObjectId like things
to check for the same value.  What's really bad however is the
simple fact that in most cases the values are going to be identical
(true result) so we do a test on every single byte in the buffer.
This prevents any benefit gained from aborting when the first few
bytes differ.

This horribly ugly unrolling shaves almost 5% off the running time
of a path limited revision walk so its just about worth doing.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>