Allow revwalk and treewalk to directly access cached object data
Its much faster to parse the cached buffer than it is to copy the
cached buffer into a new byte[], then parse that new copied array.
Less garbage is generated, so lower GC overheads, plus we get a
lot better cache locality as we are working on the buffer we did
the decompression into.

The performance difference is signficant, especially if we walk a
lot of trees at once during a revision walk.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>