New revision walking library for jgit
This is a new revision walking library that relies upon bit flags and
a per-walker object map.  The walker itself is not thread safe but
multiple walkers can be opened on the same repository at the same
time, as they share different object maps.

This design makes it easier to reset the revision walking machinery
and start a different walk, as application code can just throw away
the current RevWalk instance and create a new one.  A reset method is
also offered to clear the flags, but retain the parsed object map.

On Mac OS X PPC jgit is running about 2x slower than C Git for doing a
simple "rev-list HEAD >file" operation, but on Windows x86 we are
running on par with a Cygwin based C Git build.

Our RevWalk is sometime producing the commits in a slightly different
order from C Git, but the same commits do eventually get displayed.
The ordering difference appears to be related to not doing a
topological sort within RevWalk, as presently we are strictly obeying
only the commit time within our walk queue.

This library is not yet complete as there are many operations we need
to support for application level code, such as merge base computation,
left-right output, uninteresting nodes, object traversals, author and
message grep filtering, etc.

If we catch a checked exception during an Iterator based walk we need
to throw it back up as a runtime exception, as that is all we can
throw from within the Iterator API.  The new RevWalkException is now
used to wrap the checked exception into something we can actually
throw, and that an application can reliably catch and handle.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>