Teach RevWalk about uninteresting commits
A common requirement for applications is to see the set of commits
that are reachable from one commit A, but that are not reachable
from a different commit B.  In C Git this is "rev-list B..A" or
"rev-list ^B A".  The commit B and all commits reachable from B
are usually called "uninteresting", as we do not want to see them
in the output of the revision walk.

We implement this feature in jgit by adding an UNINTERESTING flag
bit to each uninteresting commit and carrying it through onto the
parents of that commit.  If we pop an UNINTERESTING commit from
the pending queue we carry the flag through to its parents, make
them pending, and look for another commit from the pending queue.
This ensures we don't return an UNINTERESTING commit to the caller
but yet still carry the flag through until the merge base.

We can abort the revision walk early if all pending commits have
the UNINTERESTING flag on them.  This indicates that we will never
get another interesting commit from the pending queue, as the flag
would carry into all parent commits.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>