New tree walking library for jgit
This is a parallel tree parser that allows the caller to parse a
canonical formatted Git tree object iteratively, rather than it
buffering all data up in a Tree/TreeFile graph.  This can be a
much faster way to walk through a tree's contents.

The parallel part of the parser allows us to scan multiple trees
at the same time, joining on them by their file path names, in
the native Git tree entry sort order.  This makes it easier to do
a simple two or three way diff between tree objects, such as we
might need to do in RevWalk to simplify commits along a path.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>