Implement a tree entry filtering API for TreeWalk
This is an entry filtering API similar to the one we created for
RevCommit within a RevWalk, making it possible for application code
to craft complex tree filtering logic.

At present we only have ALL and ANY_DIFF, as well as basic boolean
operators.  The ALL filter matches everything, and is our default
filter so we can avoid null pointer checks inside of the main walk
loop.  The ANY_DIFF filter can be used to implement a diff-tree
program, as it matches only tree entries that actually have some
sort of difference in the mode or SHA-1.  Mode differences also
tell us about add/delete differences as modes for non-existant
entries are always 0 (otherwise known as FileMode.MISSING).

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>