Define a path limiter for TreeWalk to reduce results
A PathFilter can be defined on a TreeWalk to limit the results it
produces to only those that match the configured path.  More than
one path may be searched for by combining path filters through the
boolean AND, OR operators already defined for a TreeFilter as well
as through the new PathFilterGroup, which is a specialized OR.

When producing a tree difference it usually is faster to include
the path filters in front of the DIFF_ANY filter, as testing for
mode or SHA-1 change almost never finds a difference.

Paths are matched using their raw byte[]s so we can avoid expensive
string conversions during execution.  This really reduces the costs
associated with looking for tree entries of particular interest to
the caller.

We allow breaking out of a TreeWalk as soon as we are beyond the last
PathFilter's configured name, assuming a PathFilterGroup was used.
On average that should allow us to avoid scanning 1/2 of each
tree, assuming the entry of interest tends towards the middle of
the listing.

If a path filter instance is matching a path name with '/' in it
then it probably needs to enter into the subtree in order to make
sure the include/exclude decision matches only the parts of the
subtree that it is actually interested in.  We delegate this decision
down through the TreeFilter graph to each of the individual filter
implementations, offering a method for the application to quickly
determine if it should enable recursive on the TreeWalk.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>