Implement path limited revision walking and parent rewriting
One way applications reduce the size of a project for either log
display or visualization is to run `git rev-list $head -- $path`,
where C Git cuts the commit DAG down to only those commits that
appear to affect the listed path(s).  For each commit walked the
rev-list process performs `git diff-tree -r $c^1 $c`, looking at
the result to see if $path appears.

We implement this same sort of concept in jgit by attaching a
TreeWalk into the middle of our RevWalk.  Commits are scanned in
two passes.

The first pass is an extended version of AbstractPendingGenerator.
Prior to enqueuing the parents of each commit we difference the
commit against its parents to see if the commit is "interesting"
(affects the paths).  If no tree entries match the TreeFilter we
likely do not want this commit to appear in the resulting DAG,
so we color with "REWRITE".  Commits are also colored with the
"UNINTERESTING" color if any of the roots were initially set that
way by the application.

All commits that were interesting to our TreeFilter are buffered
into an output list.  This ensures we are able to color all possible
parent commits with "REWRITE" before we attempt to rewrite any of
the child commits and delete/replace their parents, to make the
DAG dense and reflect only the commits the TreeFilter was matching.

During the second pass we spin out the buffer of the interesting
commits.  For each parent that is colored with "REWRITE" we do
a replacement of that parent with its own first parent (aka the
grandparent).  This replacement continues in a loop until the new
grandparent is not colored "REWRITE", was colored "UNINTERESTING",
or was an interesting merge that we could not color either way
during the first pass.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>