Add unpop support to FIFORevQueue to insert items at the front
Some applications may find it useful to be able to insert a handful of
commits into the front of a FIFORevQueue, such as if they popped the
commit, looked at it, and need to push it back into the queue to delay
it until the next pop call.

If we have no block available at the head to take the commit we insert
it into the middle of a new block, or at the end of a new block.  This
gives us the largest possible space for new entries in either direction.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>