Refactor StartGenerator to build the graph based on output behavior
Different generator implementations produce different behaviors on
their output commits.  We can control how our graph of generators
is created by looking at what flags exist on each generator.

Currently this only allows us to indicate the sort ordering of a
generator (which we only have date based) and presence of the core
REWRITE flag on commit nodes, needing a RewriteGenerator to finish
the graph before the application sees it.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>