Implement topological sorting in RevWalk
Our implementation of topological sort is almost an in-place sort.  The
algorithm is pretty simple.  As each commit gets received from the source
generator we increment the parent's in degree count, remembering the total
number of children each commit has.  During output of a commit we delay
the output if the in degree is not 0, as this indicates the commit is
ahead of at least one of its children.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>