Have RevObject extend ObjectId directly and further reduce memory
Holding a reference to an ObjectId within our RevObject costs us
approximately 12-16 bytes of memory per RevObject instance.  These
costs are from the 4 byte ObjectId reference inside of the RevObject
(assuming a 32 bit JVM) and the 8 byte object header associated with
the ObjectId we refer to.  The JVM may also be consuming an extra 4
bytes per ObjectId to round out the allocations to ensure perfect 8
byte alignment for longs and doubles.

Making RevObject extend ObjectId allows us to remove this almost 16
byte penalty.  If the JVM was rounding out the ObjectId we can now fit
our "RevObject.flags" field into the gap.  Removing the reference and
the object header provides for a significant savings per instance.

With this change our RevObject is about 8 bytes larger than C Git's
equivalent of the same structure.  There really is nothing we can do
about this extra overhead, as the JVM requires us to have it.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>